test: address native session split review feedback
